Renamed User to NewHire and built bare bones of class
