--- src/main/java/com/pjohnson_wtc/email_admin/User.java
@@ -1,49 +1,0 @@
-package com.pjohnson_wtc.email_admin;
-
-public class User {
-	private String firstName;
-	private String lastName;
-	private String department;
-	private String email;
-	private String alternateEmail;
-	
-	public User(String firstName, String lastName) {
-		this.firstName = firstName;
-		this.lastName = lastName;
-	}
-	
-	public String getFirstName() {
-		return firstName;
-	}
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-	
-	public String getLastName() {
-		return lastName;
-	}
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-	
-	public String getDepartment() {
-		return department;
-	}
-	public void setDepartment(String department) {
-		this.department = department;
-	}
-	
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	
-	public String getAlternateEmail() {
-		return alternateEmail;
-	}
-	public void setAlternateEmail(String alternateEmail) {
-		this.alternateEmail = alternateEmail;
-	}
-}
